Deprecate `#[column_name(foo)]` in favor of `#[column_name = "foo"]`
The old form was a relic of the fact that we needed to use a
`macro_rules!` macro to parse the struct at one point in the past, and
there's no way to convert a string to an ident in non-procedural macros.
The parens form is inconsistent with the rest of Diesel, such as
`#[table_name = "foo"]` and `#[sql_type = "Integer"]`.

The only attribute which still uses the parens form is `#[primary_key]`,
since it needs to be able to take multiple values.