Implement `or_filter`
I'm still not sold on this method name, but I haven't thought of a
better one. I'm open to any suggestions on better names.

This method behaves identically to `filter`, but it appends to existing
queries using `.or` instead of `.and`. This is useful for cases where
you have a large query being constructed dynamically, but you always
want some other condition to be included. A motivating use case in
crates.io is that we want to be able to do
`.or_filter(has_name(q_string))` in crates::index,
so that crates with an exact match are
always included regardless of what else we've done with the query (the
function itself is over a hundred lines of query construction).