Remove `primitive_impls!` in favor of a derive.
This adds `#[derive(SqlType)]`, which is responsible for implementing
`NotNull`, `SingleValue`, and potentially also implementing `HasSqlType`
on one or more backends. The body of `HasSqlType` is specified through
additional attributes. The only one of note is the `postgres` attribute,
which can do either static OIDs or dynamic type lookup.

`Array` and `Range` are not able to make full use of this derive, since
`Range` has a bunch of static OIDs and `Array` has a unique
implementation of `HasSqlType`.

At this point, the `custom_types` test is about as small as it's ever
going to get. The only way we could possibly eliminate any more code
there is by making it so you don't have to write as much of the
signature.