Compile successfully when `infer_schema!` is used with an empty database
If you run `infer_schema!` on an empty database, we generate this code:

    allow_tables_to_appear_in_same_query!();

That macro doesn't allow an empty token list, so we get the error
"unexpected end of macro invocation" with no indication to the user of
how to fix it.

We could just skip generating that macro in `infer_schema!`, but this is
less code.