Remove most of `InsertValues`
This removes every method on `InsertValues` except for `column_names` on
`InsertValues`. These methods were redundant with the versions on
`QueryFragment`, and we can just directly use those instead.

I'd like to actually get rid of this trait completely, but that is
impossible without a breaking change to the API of `Insertable`. As I
suspected, our implementation of `is_noop` doesn't qutie work here. The
issue is that `walk_ast` always outputs at least `()`, which is not
"nothing". I've basically added a hacky workaround to `Grouped` in the
short term (and added a test demonstrating the problem, we didn't
previously have one).

I think this hack and problem both generally go away by introducing a
`ValuesClause` struct which is responsible for both the `VALUES`
keyword, and the parenthesis. (And I guess probably the column list,
too). All of this is required if we don't want to completely
re-implement `InsertStatement` (and more importantly PG upsert,
`insert_or_ignore`, etc) for to insert from a select statement.