Partially fix our test suite on beta/nightly
Right now on beta the behavior of `file!` (or more specifically, the
cwd when compiling the crate) has changed when running `cargo test`.
It's unclear whether that change will be reverted or not, but it does
point out how brittle these tests are. We want an absolute path, not a
relative one, so using `CARGO_MANIFEST_DIR` instead seems like the way
to go.