Rewrite `AsChangeset` using syn 0.12
In order to update to syn 0.12, we're going to have to rewrite a lot of
`diesel_derives`. A lot of code is incompatible with no obvious
migration path. I also want to take this as an opportunity to do a lot
of long needed refactoring to this library.

Despite continuous assertions that outputting `compile_error!` somehow
makes our error messages significantly better (it really doesn't, it
just avoids "proc-macro panicked") -- I am panicking in error cases
here. One of the benefits of syn 0.12 is that it retains span
information, which we can use on nightly to output errors with real
diagnostics. I haven't made that jump here quite yet, but I plan to in a
followup PR.

Since this is the first derive moved over to the new crate, it includes
much more code than PRs for the rest will.