Add `.then_order_by`, which appends rather than replaces
I was originally only going to implement this for boxed queries, but I
don't think there's any reason we need to do that here. My original
reasoning was I didn't want to write code that appended tuples, but
there's actually no difference in behavior between `.order((foo, bar,
baz))` and `.order(((foo, bar), baz))` (which is the type we get from
this).

Fixes #1311.