Add `DeleteStatement::into_boxed`.
This is one half of #1128. This method (and the added type) correspond
to `SelectStatement::into_boxed` and `BoxedSelecStatement`. I had
originally stated that I thought we could just box the where clause by
default, but we could no longer do that due to backwards compatibility.
Neither of those statements are actually true.

While we can't remove the type parameter for the where clause, it is
documented that you are not allowed to rely on the types that parameter
can hold, or make any assumptions about them. So we could either change
it to be an unused parameter, or enforce that it is always a boxed where
clause.

However, we can't just change it to be unused, as we need to capture a
lifetime somewhere. So if we wanted to re-use `DeleteStatement` for
this, we'd have to go the second route of having that parameter always
be a boxed where clause. I realize now that we don't actually want to do
that though, as it would require `delete` to take the database the query
would be run against.

There's more duplicated code than I'd like (basically all of this code
was copied from either `DeleteStatement` or `BoxedSelectStatement`, but
I'm not sure that there's a ton we can do about this. It's unfortunate
that we basically have to remember to update `BoxedDeleteStatement` any
time we do anything to `DeleteStatement`, but that's no different than
the situation with selects.

While conditionally modifying the where clause of a delete statement is
definitely going to be less common than doing it for a select, I do
think that this is a reasonable feature, and one worth having.