Fix `SelectStatement` being incorrectly allowed in an expression context
This unfortunately requires a breaking change, but it is in order to fix
a bug. Any code that was relying on this impl would have generated an
invalid query, or be subject to other bugs. Additionally, all types
affected by this are internal API. Finally, the SQL types are no longer
changed by enabling the "postgres" feature, and we can start relying on
`SelectStatement` always doing what we want (there are some hacks
working around that which I have not removed in this commit).

In order to continue to be able to at least somewhat abstract over how
to get the SQL type of a select clause, I've introduced `SelectQuery`
into the public API. This trait is basically `Expression`, but only
implemented for select statements.