Allow (some) subselects to appear in an expression context
A subselect can be used as an expression by wrapping it in parenthesis
as long as the query returns exactly one column, and exactly zero or one
rows. The first constraint is trivial for us to enforce, but the only
way we can enforce the second is by automatically adding `LIMIT 1` to
the query when this method is called. We also need to make the SQL type
nullable, for the same reasons as max, min, avg, etc. They all return
null if the query returns zero rows.

If disjointness on associated types ever lands, I suspect we can just
implement `AsExpression` directly, and deprecate this method. That
said, given that the SQL type has to become nullable, it may be better
to keep the explicit method, so that we don't get questions on why
`id.eq(subselect)` isn't working.

Fixes #1308.