Replace most raw pointer usage with `NonNull`
This newly stabilized feature indicates that a pointer cannot be null.
This allows certain compiler optimizations, including but not limited to
the "null pointer optimization", which means that `Option<NonNull<T>>`
has the same size as `NonNull<T>`.

Using `NonNull` over raw pointers as much as possible in our code
potentially enables additional compiler optimizations, and better
expresses intent in several places.

Since this type was only stabilized in Rust 1.25, I have stuck the use
of the real thing behind the unstable flag, and provided a polyfill
by default. This keeps with our policy of "only bump the minimum Rust
version if you have a good reason". I don't think this is a good enough
reason to bump on its own. The polyfill lets us structure our code the
same regardless, and it should be trivial to switch to always using the
real thing once we no longer support Rust 1.24 and earlier.