Add partial support for `Numeric` on SQLite
Even though we can't support it as a proper type on that backend (there
is no native type to represent it), we still need this `FromSql` impl.
The way our query builder is structured, we can't vary the return type
of functions like `avg` based on the backend being used.

Even though this is subject to 64 bit precision, I opted to implement
support for `BigDecimal` instead of `f64` for consistency with the other
backends.

Fixes #1409.