Release v1.2.1
During the release of 1.2.0, a bug in crates.io was discovered that
prevented us from having the feature name `32-column-tables`. To work
around this, we put an `x` in front of all the affected feature names.
The bug in crates.io has now been fixed, so this `x` is no longer
necessary.